dom/tests/mochitest/integer-gemm/test_unavailable_for_webpage.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1746631
-->
<head>
  <meta charset="utf-8">
  <title>Test for Mozilla integer gemm (1746631) -- Mozilla integer gemm shouldn't be available for web pages</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1746631">Feature Test 1746631</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script type="text/javascript">
  SimpleTest.waitForExplicitFinish();

  const gemm = "mozIntGemm";
  is(gemm in WebAssembly, false, `"WebAssembly.${gemm}" shouldn't be defined for web pages`);

  SimpleTest.finish();
</script>
</pre>
</body>
</html>
